fix: remove node buffers
This module didn't have a lot of buffer use in the first place,
but we remove any references to buffers in favour of Uint8Arrays.

When quering, values come back as Buffers which are Uint8Arrays so
the conversion was unecessary.

Depends on:

- [ ] https://github.com/ipfs/interface-datastore/pull/43
- [ ] https://github.com/ipfs/js-datastore-core/pull/27

BREAKING CHANGE: remove node buffers in favour of Uint8Arrays